Выполнил д/з 5
