+++ js/upload-photo.js
@@ -1,0 +1,1 @@
+import './edit-photo.js';
